+++ imports/startup/client/lib/sharedFunctions.js
@@ -1,5 +1,5 @@
 // Renvoi a Number if the field value is a number otherwise return NaN
-filterInt = function(value) {
+export const filterInt = function(value) {
 	if (/^(\-|\+)?([0-9]+|Infinity)$/.test(value)) {
 		return Number(value);
 	} else {
@@ -7,7 +7,7 @@ filterInt = function(value) {
 	}
 };
 
-isValidUrl = function(url) {
+export const isValidUrl = function(url) {
 	var urlregex = new RegExp('^(http|https|ftp)\://([a-zA-Z0-9\.\-]+(\:[a-zA-Z0-9\.&amp;%\$\-]+)*@)*((25[0-5]|2[0-4][0-9]|[0-1]{1}[0-9]{2}|[1-9]{1}[0-9]{1}|[1-9])\.(25[0-5]|2[0-4][0-9]|[0-1]{1}[0-9]{2}|[1-9]{1}[0-9]{1}|[1-9]|0)\.(25[0-5]|2[0-4][0-9]|[0-1]{1}[0-9]{2}|[1-9]{1}[0-9]{1}|[1-9]|0)\.(25[0-5]|2[0-4][0-9]|[0-1]{1}[0-9]{2}|[1-9]{1}[0-9]{1}|[0-9])|([a-zA-Z0-9\-]+\.)*[a-zA-Z0-9\-]+\.(com|edu|gov|int|mil|net|org|biz|arpa|info|name|pro|aero|coop|museum|[a-zA-Z]{2}))(\:[0-9]+)*(/($|[a-zA-Z0-9\.\,\?\'\\\+&amp;%\$#\=~_\-]+))*$');
 	if (urlregex.test(url)) {
 		return url;
@@ -16,7 +16,7 @@ isValidUrl = function(url) {
 	}
 };
 
-getFullName = function(obj) {
+export const getFullName = function(obj) {
 	var fullName = '';
 	if (obj.firstName === undefined && obj.lastName === undefined) {
 		return fullName;
@@ -32,7 +32,7 @@ getFullName = function(obj) {
 	}
 };
 
-getAuthors = function(list) {
+export const getAuthors = function(list) {
 	var author = '';
 	if (list.length === 0) {
 		return author;
