+++ imports/startup/server/index.js
@@ -1,1 +1,2 @@
 import './kadira.js';
+import './registerApi.js';
