+++ imports/startup/server/registerApi.js
@@ -1,0 +1,11 @@
+// Collections
+import '/imports/api/mangasData/schema.js';
+import '/imports/api/mangas/schema.js';
+
+// Publications
+import '/imports/api/mangasData/server/publications.js';
+import '/imports/api/mangas/server/publications.js';
+
+// Methods
+import '/imports/api/mangasData/methods.js';
+import '/imports/api/mangas/methods.js';
