--- imports/ui/components/addMangaToWaitingList/addMangaToWaitingList.jade
@@ -1,11 +1,0 @@
-template(name="addMangaToWaitingList")
-	.panel.panel-default
-		.panel-heading
-			h2.panel-title {{_ "dontSeeMangasYouOwn"}}
-		.panel-body
-			form.form-inline
-				.form-group
-					label(for="mangasToAdd") {{_ "mangasName"}}
-					input.form-control#mangasNameToAdd(type="text")
-					button.btn.btn-success#addMangaToWaitingList(type="button") {{_ "add"}}
-			p {{_ "waitingExplanation"}}
