+++ imports/ui/components/mangasCoverList/mangasCoverList.jade
@@ -2,5 +2,3 @@ template(name="mangasCoverList")
 	#coverList.grid
 		each covers
 			+mangaCover
-	if currentUser
-		+addMangaToWaitingList
