+++ imports/ui/pages/addMangaForUser/step1.jade
@@ -1,4 +1,5 @@
 template(name="addMangaForUserStep1")
+	+header
 	.panel.panel-default
 		.panel-heading
 			h2 {{_ "whichMangaToAdd"}}
