--- imports/ui/pages/addMangaForUser/step1/addMangaForUserStep1.js
@@ -1,7 +1,0 @@
-Template.addMangaForUserStep1.events({
-	'click .mangas-cover': function() {
-		return Router.go('addMangaForUserStep2', {
-			_id: this._id
-		});
-	}
-});
