+++ server/main.js
@@ -1,3 +1,1 @@
 import '/imports/startup/server';
-import '/imports/api/mangas/schema.js';
-import '/imports/api/mangas/server/publications.js';
